Tweaking
Reduced glitchyness by making sure that the random intersection values
are taken from a noise space to be reproducible for both edges for the
points involved. Then we invert the result based on atan2 to get the
opposite on both sides/depending on in which direction we traverse the
coordinates.

Finally find a better limit for the number of possible subdivision steps
by using the log2 of the minimum height of the triangle. This both makes
sure we have the right subdivision steps for the general scale of the
triangle, it also presents super thin triangles from being subdivided a
gazillion times.